+++ core/IGenerator.h
@@ -9,11 +9,12 @@
 #define IGenerator_h
 
 #include "Parameters.h"
+#include <algorithm>
 
 class IGenerator {
 public:
     virtual void initialize() = 0;
-    virtual void activateSlice(long onset, long offset, long length, bool reset) = 0;
+    virtual void activateSlice(uint32_t onset, uint32_t offset, uint32_t length, bool reset) = 0;
     virtual void generate(float* out0, float* out1) = 0;
     virtual void setDirection(vlly::spotykach::Direction direction) = 0;
     virtual void reset() = 0;
